requirements.txt
streamlit
numpy
math
matplotlib
streamlit-matplotlib
pillow